changed price to float to easly deal with data

+++ main.py
@@ -4,4 +4,4 @@ import pandas as pd
 df = pd.read_csv('bestsellers.csv')
 df.drop_duplicates(inplace=True)
 df.rename(columns={"Name": "Title", "Year": "Publication", "User Rating": "Rating"}, inplace=True)
-
+df["Price"] = df["Price"].astype(float)
